chore: refactor `anvil` for library use (#10130)
* make `anvil` easier accessible as a library

* clean up

* doc nits
